requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
python-keycloak>=3.0
django-cors-headers>=4.0
gunicorn>=21.2
python-dotenv>=1.0
requests>=2.31
